Move constructor rules tests to their own source file
The nature and behavior of decoration constructors is important enough to merit having these behaviors tested in their own category.  Issue #500 specifically will require just such a class of test.
